Добавлены улучшения в обработку ошибок и поддержку дисплея для FakeTech V5
- Реализована обработка ошибок для события `ble_ota_unsupported` в OTA процессе.
- Обновлена документация API с новыми кодами валидации команд для BLE.
- Добавлены зависимости для работы с OLED дисплеем (Adafruit SSD1306 и GFX).
- Оптимизированы функции отображения сообщений на дисплее и добавлена поддержка самотестирования.

+++ firmware/src/faketec/x25519_keys_nrf.cpp
@@ -82,7 +82,7 @@ static std::atomic<uint32_t> s_throttleSeq{0};  // even=stable, odd=writer
 /** Слот троттла по хэшу peerId — не round-robin: иначе 4 чужие TX затирают lastSend и дебаунс HELLO+KEY ломается */
 static int throttleSlotForPeer(const uint8_t* peerId) {
   uint32_t h = 2166136261u;
-  for (int i = 0; i < protocol::NODE_ID_LEN; i++) {
+  for (size_t i = 0; i < protocol::NODE_ID_LEN; i++) {
     h ^= peerId[i];
     h *= 16777619u;
   }
